requirements.txt
requests>=2.31.0,<2.32
feedparser>=6.0.10,<6.1
discord-webhook>=1.2.1,<1.3
beautifulsoup4>=4.12.2,<4.13
python-dotenv==1.0.0
schedule==1.2.0